Move supabase location
