+++ src/utils/supabase/static-props.js
